requirements.txt
aws-cdk-lib>=2.121.0
constructs>=10.0.0